+++ src/main.rs
@@ -1,33 +1,22 @@
-use actix_files::NamedFile;
-use actix_governor::{Governor, GovernorConfigBuilder};
-use actix_web::{
-    http::StatusCode,
-    middleware::{self, ErrorHandlers},
-    web, App, HttpResponse, HttpServer,
-};
-use anyhow::{Context, Result};
-use std::{
-    env,
-    path::{Path, PathBuf},
-    sync::OnceLock,
+// no main needed for non-ssr because
+// wasm is loaded using `hydrate()` from lib
+#![cfg_attr(not(feature = "ssr"), no_main)]
+#![cfg(feature = "ssr")]
+
+use {
+    actix_files::NamedFile,
+    actix_governor::{Governor, GovernorConfigBuilder},
+    actix_web::{middleware, web, App, HttpResponse, HttpServer},
+    anyhow::{Context, Result},
+    leptos::*,
+    leptos_actix::LeptosRoutes,
+    std::{env, path::Path},
 };
 
 mod catalog;
 mod logging;
-mod pages;
 
 type Date = chrono::NaiveDate;
-
-const ORG_NAME: &str = "Cap Hill Rust";
-const MEETUP_URL: &str = "https://www.meetup.com/Cap-Hill-Rust/";
-const GITHUB_URL: &str = "https://github.com/JesusGuzmanJr/cap-hill-rust";
-
-static ASSETS: OnceLock<PathBuf> = OnceLock::new();
-
-#[inline]
-fn assets() -> &'static Path {
-    ASSETS.get().expect("not initialized")
-}
 
 #[actix_web::main]
 async fn main() -> Result<()> {
@@ -36,9 +25,24 @@ async fn main() -> Result<()> {
 
     let bind_address = env::var("BIND_ADDRESS").with_context(|| "BIND_ADDRESS is not set")?;
 
-    ASSETS
-        .set(Path::new(&env::var("ASSETS_DIR").with_context(|| "ASSETS_DIR is not set")?).into())
-        .expect("already initialized");
+    let leptos_options = {
+        LeptosOptions {
+            env: if cfg!(debug_assertions) {
+                leptos_config::Env::DEV
+            } else {
+                leptos_config::Env::PROD
+            },
+            site_root: env::var("ASSETS").with_context(|| "ASSETS is not set")?,
+            ..leptos_config::get_config_from_env()
+                .expect("failed to get leptos config")
+                .leptos_options
+        }
+    };
+
+    // Generate the list of routes in your Leptos App
+    let routes = leptos_actix::generate_route_list(cap_hill_rust::App);
+
+    let leptos_app_data = web::Data::new(leptos_options.to_owned());
 
     HttpServer::new(move || {
         const FAVICON: &str = "favicon.ico";
@@ -49,37 +53,41 @@ async fn main() -> Result<()> {
                 "/health",
                 web::get().to(|| async { HttpResponse::NoContent().finish() }),
             )
-            .service(
-                web::scope("")
-                    .wrap(
-                        ErrorHandlers::new()
-                            .handler(StatusCode::NOT_FOUND, pages::not_found::handler),
+            .route(
+                FAVICON,
+                web::get().to(|data: web::Data<LeptosOptions>| async move {
+                    NamedFile::open_async(
+                        Path::new(&data.get_ref().site_root)
+                            .join("favicons")
+                            .join("favicon.ico"),
                     )
-                    .route(
-                        FAVICON,
-                        web::get().to(move || async {
-                            NamedFile::open_async(assets().join("favicons").join("favicon.ico"))
-                                .await
-                        }),
-                    )
-                    .route(
-                        ROBOTS,
-                        web::get().to(move || async {
-                            NamedFile::open_async(assets().join(ROBOTS)).await
-                        }),
-                    )
-                    .service(pages::index::handler)
-                    .service(pages::library::handler)
-                    .service(catalog::get_catalog)
-                    .service(actix_files::Files::new("/assets", &assets()))
-                    .wrap(actix_web::middleware::Logger::new("%s for %r %a in %Ts"))
-                    .wrap(middleware::Condition::new(
-                        cfg!(not(debug_assertions)),
-                        actix_web_lab::middleware::RedirectHttps::with_hsts(
-                            actix_web_lab::header::StrictTransportSecurity::recommended(),
-                        ),
-                    )),
+                    .await
+                }),
             )
+            .route(
+                ROBOTS,
+                web::get().to(|data: web::Data<LeptosOptions>| async move {
+                    NamedFile::open_async(Path::new(&data.get_ref().site_root).join(ROBOTS)).await
+                }),
+            )
+            .leptos_routes(
+                leptos_options.to_owned(),
+                routes.to_owned(),
+                cap_hill_rust::App,
+            )
+            .service(actix_files::Files::new("/", &leptos_options.site_root).use_hidden_files())
+            .app_data(leptos_app_data.clone())
+            .service(actix_files::Files::new(
+                "/assets",
+                leptos_options.site_root.clone(),
+            ))
+            .wrap(actix_web::middleware::Logger::new("%s for %r %a in %Ts"))
+            .wrap(middleware::Condition::new(
+                cfg!(not(debug_assertions)),
+                actix_web_lab::middleware::RedirectHttps::with_hsts(
+                    actix_web_lab::header::StrictTransportSecurity::recommended(),
+                ),
+            ))
             .wrap(Governor::new(
                 &GovernorConfigBuilder::default()
                     .finish()